making sure all changes are up to date
