+++ src/App/Info/MapCard/mapIcon.tsx
@@ -1,0 +1,8 @@
+import L from "leaflet";
+
+export const customIcon = new L.Icon({
+  iconUrl: "/Photos/MapCardPhotos/marker-icon-2x.png",
+  iconSize: [15, 20],
+  iconAnchor: [8, 32],
+  popupAnchor: [0, -32],
+});
